Move `use-mock-datastore.ts` to `datastore` folder
