+++ packages/mock-block-dock/src/data/index.ts
@@ -1,4 +1,4 @@
-import { MockData } from "../use-mock-block-props/use-mock-datastore";
+import { MockData } from "../datastore/use-mock-datastore";
 import { entities } from "./entities";
 import { entityTypes } from "./entity-types";
 import { linkedAggregationDefinitions } from "./linked-aggregation-definitions";
