+++ packages/mock-block-dock/src/mock-block-dock-context.tsx
@@ -15,7 +15,7 @@ import {
   useState,
 } from "react";
 
-import { MockData } from "./use-mock-block-props/use-mock-datastore";
+import { MockData } from "./datastore/use-mock-datastore";
 
 type MockBlockDockInfo = {
   blockEntity: Entity;
